+++ addon/genmqtt.py
@@ -68,6 +68,7 @@ class MyGenPush(MySupport):
         blacklist=None,
         flush_interval=float("inf"),
         use_numeric=False,
+        use_numeric_object=False,
         strlist_json = False,
         debug=False,
         loglocation=ProgramDefaults.LogPath,
@@ -78,6 +79,7 @@ class MyGenPush(MySupport):
         self.Callback = callback
 
         self.UseNumeric = use_numeric
+        self.UseNumericObject = use_numeric_object
         self.StrListJson = strlist_json
         self.debug = debug
         self.Exiting = False
@@ -211,7 +213,7 @@ class MyGenPush(MySupport):
         while True:
             try:
 
-                if not self.UseNumeric:
+                if not self.UseNumeric and not self.UseNumericObject:
                     statusdata = self.SendCommand("generator: status_json")
                     maintdata = self.SendCommand("generator: maint_json")
                     outagedata = self.SendCommand("generator: outage_json")
@@ -259,7 +261,10 @@ class MyGenPush(MySupport):
             for key, item in node.items():
                 if isinstance(item, dict):
                     CurrentPath = PathPrefix + "/" + str(key)
-                    self.CheckDictForChanges(item, CurrentPath)
+                    if self.UseNumericObject and self.DictIsTopicJSON(item):
+                        self.CheckForChanges(CurrentPath, json.dumps(item, sort_keys=False))
+                    else:
+                        self.CheckDictForChanges(item, CurrentPath)
                 elif isinstance(item, list):
                     CurrentPath = PathPrefix + "/" + str(key)
                     if self.ListIsStrings(item):
@@ -271,7 +276,10 @@ class MyGenPush(MySupport):
                     else:
                         for listitem in item:
                             if isinstance(listitem, dict):
-                                self.CheckDictForChanges(listitem, CurrentPath)
+                                if self.UseNumericObject and self.DictIsTopicJSON(item):
+                                    self.CheckForChanges(CurrentPath, json.dumps(item, sort_keys=False))
+                                else:
+                                    self.CheckDictForChanges(listitem, CurrentPath)
                             else:
                                 self.LogError(
                                     "Invalid type in CheckDictForChanges: %s %s (2)"
@@ -283,6 +291,17 @@ class MyGenPush(MySupport):
         else:
             self.LogError("Invalid type in CheckDictForChanges %s " % str(type(node)))
 
+    # ---------- MyGenPush::ListIsStrings--------------------------------------- 
+    def DictIsTopicJSON(self, entry):
+        try:
+            if not isinstance(entry, dict):
+                return False
+            if "type" in entry.keys() and "value" in entry.keys() and "unit" in entry.keys():
+                return True
+            return False
+        except Exception as e1:
+            self.LogErrorLine("Error in DictIsTopicJSON: " + str(e1))
+            return False
     # ---------- MyGenPush::ListIsStrings---------------------------------------
     # return true if every element of list is a string
     def ListIsStrings(self, listinput):
@@ -403,6 +422,9 @@ class MyMQTT(MyCommon):
             )
             self.UseNumeric = config.ReadValue(
                 "numeric_json", return_type=bool, default=False
+            )
+            self.UseNumericObject = config.ReadValue(
+                "numeric_json_object", return_type=bool, default=False
             )
             self.StringListJson = config.ReadValue(
                 "strlist_json", return_type=bool, default=False
@@ -550,6 +572,7 @@ class MyMQTT(MyCommon):
                 blacklist=self.BlackList,
                 flush_interval=self.FlushInterval,
                 use_numeric=self.UseNumeric,
+                use_numeric_object=self.UseNumericObject,
                 strlist_json = self.StringListJson,
                 debug=self.debug,
                 port=port,