+++ changelog.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - More updates for ComAp controller
 - Update to gentankutil to add option to check battery and missed readings. See Add On page to enable this functionality.
 - Minor update to error handling of fuel gauge 
+- Added new option to return JSON for numerics on MQTT addon genmqtt.py
 
 ## V1.19.00 -2023-11-09
 - Updates to genmonmain.sh, startgenmon.sh and genloader.sh for Debian bookworm.